+++ src/components/empty.tsx
@@ -14,6 +14,7 @@ import {
   Paper,
 } from "@mui/material";
 import { Link } from "react-router-dom";
+import MetahkgLogo from "./logo";
 /*
  * just a template for large screens if there's no content
  * e.g. /category/:id, in which there's no main content but only the menu
@@ -35,13 +36,7 @@ export default function Empty() {
       >
         <div style={{ margin: "50px", width: "100%" }}>
           <div style={{ display: "flex", alignItems: "center" }}>
-            <img
-              className="svgwhite"
-              height="40px"
-              width="50px"
-              src="/logo.svg"
-              alt="Metahkg"
-            />
+            <MetahkgLogo height={40} width={50} svg light/>
             <h1 style={{ color: "white" }}>Metahkg</h1>
           </div>
           <List>